netcom-web: new management of HTTP passwords
We now have a file in /etc/conf.d/, called http-passwords, to store
the http passwords. A script, executed at boot time and when passwords
are changed from the Web interface, is executed to update the HTTPd
configuration file according to the new passwords. This allows to keep
the HTTP passwords intact accross firmware upgrades.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>