netcom: Add usr/share/release/variables file to the filesystem
The usr/share/release/variables file is added to the filesystem
automatically during the build process. Its template is stored in
board/getinge/netcom/rootfs-additions/usr/share/release/variables.

In the post-build script (executed at every build), we copy this
template to the target filesystem, and we replace the __BUILD_TIME__
and __BUILD__ magic strings by the build date/time and the build id.

The build id is generated by the new
board/getinge/netcom/get-build-id.sh script. A large comment at the
top of this file explains how it works.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>