netcom-apps: improve daemon start logic
Allow daemons to define an options="" string in their .startupdb
file. Those options are passed at daemon startup. This is needed for
sersrvd, which must be started in the background (i.e, with option "
&").

Also test the license before starting the daemon, and show "No
license" error message at daemon startup when the license isn't
available for the daemon.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>