netcom: change approach for localtime link in conf.d filesystem image
The /etc/localtime symbolic link is created by Buildroot after all
packages are created, but before the post-build script is generated.

This /etc/localtime symbolic link was moved to /etc/conf.d/ by the
post-build script, so that it would be picked up when creating the
/etc/conf.d/ filesystem image.

Unfortunately, the etc-confd package, which creates the etc/conf.d/
default filesystem image, is built as a package. This means that it is
created before the /etc/localtime symbolic link has appeared: the
default filesystem image for /etc/conf.d/ does not contain the
symbolic link.

We cannot easily move the etc-confd package logic in the post-build
script, because it needs to depend on the host-mtd package, to make
sure that mkfs.jffs2 is available on the host system.

To solve this problem, we just create the /etc/conf.d/localtime
symbolic link in the etc-confd build procedure, linked to the UTC
timezone. This means that the default timezone is hardcoded in the
etc-confd package recipe, and no longer using the corresponding
Buildroot configuration option. The post build script removes the
/etc/localtime created by Buildroot, and replaces it by a link to
/etc/conf.d/localtime.

Fixes issue 601b.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>