netcom: adjust vsftpd/telnetd handling
Instead of having custom init scripts, vsftpd and telnetd now use the
normal daemon infrastructure. The netcom-common init script is changed
to accept daemons that are not in /usr/bin, the daemon path is fetched
using the which program.

The telnetd and vsftpd daemons are however not made visible inside the
web interface, thanks to the visible=0 value in their startupdb
file. Besides that, they behave like any other daemon: a
{vsftpd,telnetd}.service file in /etc/conf.d/ containing enabled=1
will make those daemons start automatically on boot, and
/etc/init.d/S70{telnetd,vsftpd} manual-{start,stop} will manually
start and stop them.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>