package: add support for the wf111 WiFi driver and its utilities
Adds support for the BlueGiga WF111 WiFi driver and the binary utilities
distributed alongside the driver. An account is required to download the
sources from the BlueGiga website, which can be created freely. The
driver is available for armv5, arvmv7a and i386.

It is not possible to automatically retrieve the sources, because of the
required user account needed on the BlueGiga website, an option is added
to let the Buildroot user specify the directory were the driver's
tarball was downloaded.

Finally, two options must be selected in the Linux kernel configuration:
CONFIG_WIRELESS_EXT and CONFIG_WEXT_PRIV.

Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>