package/glibc: bump to 2.39
The 3 patches we had have all been merged upstream as of 2.39, so they
can be dropped.

The libcrypt library is gone from glibc, and therefore the
--enable-crypt option is also gone.

The LICENSES file has changed with the addition of a copyright notice
from IBM (some files contributed by IBM have been relicensed), and the
license is MIT, which is already listed in GLIBC_LICENSES, so no
changes are needed on our side.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>