package/gdb: add support for GDB 15.1
As detailed in the release notes [1]:

  Building GDB and GDBserver now requires a C++17 compiler (for
  instance, GCC 9 or later).

Therefore, we make the option BR2_GDB_VERSION_15 depend on both
BR2_HOST_GCC_AT_LEAST_9 *and* BR2_TOOLCHAIN_GCC_AT_LEAST_9. One might
wonder why on both. It's because BR2_GDB_VERSION_15 defines both the
version of gdb used for the host *and* target, because we want them to
be in sync (to make sure our gdbserver is in sync with host-gdb).

The patches from gdb 14.1 are all ported to 15.1, with some minor
adaptation to 2 of them.

[1] https://lwn.net/Articles/981080/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Romain Naour <romain.naour@smile.fr>