configs/qemu_sh4*_rd2: use an external initramfs
In commit
44a6b7866c95ca5459a53a50eab5ab1d38b6b8e1 ("configs/qemu_sh4*: switch
to initramfs"), we switched from generating an ext2 image of the root
filesystem to using an initramfs built into the kernel, to work around
a bug in the Qemu IDE emulation on SH4 R2D.

It turns out that using a built-in initramfs only works if the root
filesystem is really really small. Just even adding the libstdc++
library makes it large enough that the kernel image no longer boots.

As an alternative, we use an external initramfs. It also has size
limitations, as the SH4 R2D board is emulated by Qemu with only 64 MB
of RAM, but it also to go a bit beyond what is possible with a
built-in initramfs.

Suggested-by: Waldemar Brodkorb <wbx@openadk.org>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>